Move account layout to account components
